# Optimize UUID types