--- a/compare.go
+++ b/compare.go
@@ -1,6 +1,8 @@
 package parquet
 
 import (
+	"encoding/binary"
+
 	"github.com/segmentio/parquet-go/deprecated"
 )
 
@@ -90,4 +92,39 @@ func compareUint64(v1, v2 uint64) int {
 	default:
 		return 0
 	}
+}
+
+func compareBE128(v1, v2 *[16]byte) int {
+	x := binary.BigEndian.Uint64(v1[:8])
+	y := binary.BigEndian.Uint64(v2[:8])
+	switch {
+	case x < y:
+		return -1
+	case x > y:
+		return +1
+	}
+	x = binary.BigEndian.Uint64(v1[8:])
+	y = binary.BigEndian.Uint64(v2[8:])
+	switch {
+	case x < y:
+		return -1
+	case x > y:
+		return +1
+	default:
+		return 0
+	}
+}
+
+func lessBE128(v1, v2 *[16]byte) bool {
+	x := binary.BigEndian.Uint64(v1[:8])
+	y := binary.BigEndian.Uint64(v2[:8])
+	switch {
+	case x < y:
+		return true
+	case x > y:
+		return false
+	}
+	x = binary.BigEndian.Uint64(v1[8:])
+	y = binary.BigEndian.Uint64(v2[8:])
+	return x < y
 }