--- a/dictionary_amd64.go
+++ b/dictionary_amd64.go
@@ -4,6 +4,8 @@ package parquet
 
 import (
 	"unsafe"
+
+	"github.com/segmentio/parquet-go/internal/bits"
 )
 
 //go:noescape
@@ -35,6 +37,9 @@ func dictionaryLookupByteArrayString(dict []uint32, page []byte, indexes []int32
 
 //go:noescape
 func dictionaryLookupFixedLenByteArrayString(dict []byte, len int, indexes []int32, rows array, size, offset uintptr) errno
+
+//go:noescape
+func dictionaryLookupFixedLenByteArrayPointer(dict []byte, len int, indexes []int32, rows array, size, offset uintptr) errno
 
 func (d *int32Dictionary) lookup(indexes []int32, rows array, size, offset uintptr) {
 	checkLookupIndexBounds(indexes, rows)
@@ -78,6 +83,12 @@ func (d *uint32Dictionary) lookup(indexes []int32, rows array, size, offset uint
 func (d *uint64Dictionary) lookup(indexes []int32, rows array, size, offset uintptr) {
 	checkLookupIndexBounds(indexes, rows)
 	dictionaryLookup64bits(d.values, indexes, rows, size, offset).check()
+}
+
+func (d *be128Dictionary) lookupPointer(indexes []int32, rows array, size, offset uintptr) {
+	checkLookupIndexBounds(indexes, rows)
+	dict := bits.Uint128ToBytes(d.values)
+	dictionaryLookupFixedLenByteArrayPointer(dict, 16, indexes, rows, size, offset).check()
 }
 
 func (d *int32Dictionary) bounds(indexes []int32) (min, max int32) {