The code was starting to be crippled with checks for fixed-length byte arrays of size 16 as a mechanism to use better code paths for the common case of using columns of type `[16]byte` (e.g. UUIDs, Trace IDs, etc...).

Rather than having these details spread through the codebase, I would like to introduce a new set of types which provide the internal specialization for these column types.

Having knowledege of the type size enables more optimizations as well, both on the Go side (see `compareBE128`, `lessBE128`), but also in assembly routines, for example in the dictionary encoding:

```
name                                        old time/op  new time/op  delta
Dictionary/Bounds/FIXED_LEN_BYTE_ARRAY(16)  8.05µs ± 0%  1.14µs ± 0%   -85.87%  (p=0.000 n=8+10)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)  20.0µs ± 1%  20.9µs ± 1%    +4.32%  (p=0.000 n=10+9)
Dictionary/Lookup/FIXED_LEN_BYTE_ARRAY(16)   857ns ± 1%   856ns ± 0%      ~     (p=0.309 n=8+9)

name                                        old value/s  new value/s  delta
Dictionary/Bounds/FIXED_LEN_BYTE_ARRAY(16)    124M ± 0%    879M ± 0%  +607.77%  (p=0.000 n=8+10)
Dictionary/Insert/FIXED_LEN_BYTE_ARRAY(16)   50.0M ± 1%   47.9M ± 1%    -4.14%  (p=0.000 n=10+9)
Dictionary/Lookup/FIXED_LEN_BYTE_ARRAY(16)   1.17G ± 1%   1.17G ± 0%      ~     (p=0.236 n=8+9)
```

The increase on the insert path is due to the Go runtime optimizing the map implementation when the key is a `string`. The key type being changed to `[16]byte`, there is a small runtime cost. What the benchmark doesn't show tho is the fact that using `string` keys requires more heap allocations and memory copies.